Major performance optimization
Changed the outer for loop with the inner runConcurrently operator.

The previous version was:
- Outer for loop for every pixel
- Run concurrently for every transport path of that pixel

The new version is:
- Outer for loop for every transport path
- Run concurrently for every pixel of that transport path index

With this change, the processing time went to 25%!

The reasoning is that we were taking a lot of time on the main thread creating small batches of threads. Now, we take less time on the main thread by creating bigger batches of threads.

The new format allowed us to remove one intermediate buffer but we store more data in general.